pyproject.toml
[project]
name = "screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "finnhub-python>=2.4.25",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "rookiepy>=0.5.6",
    "tabulate>=0.9.0",
    "tradingview-screener>=3.0.0",
    "yfinance>=0.2.66",
]
